wip: more idiomatic code
